Fix: quita y reacomodo de campos del formulario
